+++ core/domain/src/main/java/com/jacqulin/calcalc/core/domain/model/Ingredients.kt
@@ -2,9 +2,9 @@ package com.jacqulin.calcalc.core.domain.model
 
 data class Ingredient(
     val name: String,
-    val weight: Double,
-    val calories: Double,
-    val protein: Double,
-    val fat: Double,
-    val carb: Double
+    val weight: Int,
+    val calories: Int,
+    val protein: Int,
+    val fat: Int,
+    val carb: Int
 )
